input: keyboard: imx: snvs_pwrkey: Add support for imx6qdl
On imx6qdl the interrupt only triggers on the release of the key.
Normally, the driver is looking for a change in the state of the key,
but since the interrupt triggers on release the key value is always 0,
so there was no event.

Add "key-release-only" boolean dts property to address this issue,
and create both key press and key release events when the key
is actually released.

Signed-off-by: Eran Matityahu <eran.m@variscite.com>