media: mxc: Enable dual pixel mode for YUV422
On the imx8mm the MIPI CSI-2 receiver can be configured to transfer two
pixel components of YUV422 per clock cycle by setting the PIXEL_MODE
bits in the MIPI_CSI_ISP_CONFIGn register to "Dual pixel mode".

The CSI bridge has to be configured accordingly by settings the
MIPI_DOUBLE_CMPNT bit in the CSI_CSICR18.

NOTE: We expected that setting the MIPI_DOUBLE_CMPNT bit would break
YUV422 for the imx8mq since it's using the same CSI bridge but a
different MIPI CSI-2 receiver. The CSI receiver on the imx8mq can't be
configured with regards to the way the data is transfered. For the
YUV422 8-bit format the datasheet says:

  13.8.3.6.15 YUV422 8 bit (Data Type = 0x1E)
  With YUV422 8 bit is received as UYVY sequence. With the Rx controller
  configured for single pixel mode, data_out[7:0] contains the U1
  component, bits [15:8] Y1. On the next data valid cycle bits [7:0] V1,
  bits [15:8] Y2.

So it should *always* send 16-bit data to the bridge, which should only
be processed correctly with the MIPI_DOUBLE_CMPNT set. But somehow it
doesn't seem to make a difference on the imx8mq whether this bit is set
or not