implement cache invalidation for MMAP
When the property "bsl,dma-invalidate;" is added to the device tree, the data cache area of frame buffers
will get invalidated after each DMA transfer. The property takes an optional 64 bit integer parameter.
If this parameter is given, then the invalidation is only done for buffers that are smaller than this value.

Example with parameter:
  bsl,dma-invalidate = <0 2097152>; // invalidate buffers smaller than 2MB